+++ libs/ImageProcessingTools.h
@@ -13,10 +13,10 @@
 PGM_PPM<byte> apply_mask(int(* mask)[3], byte** matrix, long nrl, long nrh, long ncl, long nch);
 PGM_PPM<byte> normeGradient(byte** sobelX, byte** sobelY, long nrl, long nrh, long ncl, long nch);
 PGM_PPM<byte> binariser(byte** matrix, long nrl, long nrh, long ncl, long nch, int seuil);
-double percentageOfContoursInImage(byte** matrix, long nrl, long nrh, long ncl, long nch); 
+float percentageOfContoursInImage(byte** matrix, long nrl, long nrh, long ncl, long nch); 
 long* histogramme(byte** matrix, long nrl, long nrh, long ncl, long nch);
 float* rateColors(rgb8** matrix, long nrl, long nrh, long ncl, long nch);
 void processImage(char* path, char* filename, char* outPath);
-double percentageOfContoursInImageTrue(byte** matrix, long nrl, long nrh, long ncl, long nch);
+float percentageOfContoursInImageTrue(byte** matrix, long nrl, long nrh, long ncl, long nch);
     
 #endif
